removed pools endpoint
