+++ packages/ui/hooks/fuse/useCrossFusePools.ts
@@ -1,15 +1,17 @@
 import { FusePoolData, SupportedChains } from '@midas-capital/types';
 import { useQueries } from '@tanstack/react-query';
-import axios from 'axios';
 import { useMemo } from 'react';
 
+import { config } from '@ui/config/index';
 import { useMultiMidas } from '@ui/context/MultiMidasContext';
 import { useUSDPrices } from '@ui/hooks/useUSDPrices';
+import { FusePoolsPerChain } from '@ui/types/ChainMetaData';
 import { Err, PoolsPerChainStatus } from '@ui/types/ComponentPropsType';
 import { MarketData, PoolData } from '@ui/types/TokensDataMap';
+import { poolSort } from '@ui/utils/sorts';
 
 export const useCrossFusePools = (chainIds: SupportedChains[]) => {
-  const { address } = useMultiMidas();
+  const { address, getSdk } = useMultiMidas();
   const { data: prices } = useUSDPrices(chainIds);
 
   const poolsQueries = useQueries({
@@ -17,64 +19,74 @@ export const useCrossFusePools = (chainIds: SupportedChains[]) => {
       return {
         queryKey: ['useCrossFusePools', chainId, address, prices && prices[chainId.toString()]],
         queryFn: async () => {
-          if (chainId && prices && prices[chainId.toString()]) {
-            try {
-              const res = await axios.post('/api/pools', {
-                chains: [chainId],
-                address,
+          const sdk = getSdk(Number(chainId));
+
+          if (chainId && prices && prices[chainId.toString()] && sdk) {
+            const chainPools: FusePoolsPerChain = {};
+            const _allPools: FusePoolData[] = [];
+
+            const pools = await sdk.fetchPoolsManual({ from: address });
+            let visiblePools: FusePoolData[] = [];
+            if (pools && pools.length !== 0) {
+              type configKey = keyof typeof config;
+
+              const hidePools = (config[`hidePools${sdk.chainId}` as configKey] as string[]) || [];
+              pools.map((pool) => {
+                if (pool && !hidePools.includes(pool.id.toString())) {
+                  visiblePools.push({ ...pool, chainId: Number(sdk.chainId) });
+                }
               });
-
-              const _allPools = res.data.allPools as FusePoolData[];
-
-              const allPools: PoolData[] = await Promise.all(
-                _allPools.map((pool) => {
-                  const assetsWithPrice: MarketData[] = [];
-                  const { assets } = pool;
-
-                  if (assets && assets.length !== 0) {
-                    assets.map((asset) => {
-                      assetsWithPrice.push({
-                        ...asset,
-                        supplyBalanceFiat:
-                          asset.supplyBalanceNative * prices[pool.chainId.toString()].value,
-                        borrowBalanceFiat:
-                          asset.borrowBalanceNative * prices[pool.chainId.toString()].value,
-                        totalSupplyFiat:
-                          asset.totalSupplyNative * prices[pool.chainId.toString()].value,
-                        totalBorrowFiat:
-                          asset.totalBorrowNative * prices[pool.chainId.toString()].value,
-                        liquidityFiat:
-                          asset.liquidityNative * prices[pool.chainId.toString()].value,
-                      });
-                    });
-                  }
-                  const adaptedFusePoolData: PoolData = {
-                    ...pool,
-                    assets: assetsWithPrice,
-                    totalLiquidityFiat:
-                      pool.totalLiquidityNative * prices[pool.chainId.toString()].value,
-                    totalAvailableLiquidityFiat:
-                      pool.totalAvailableLiquidityNative * prices[pool.chainId.toString()].value,
-                    totalSuppliedFiat:
-                      pool.totalSuppliedNative * prices[pool.chainId.toString()].value,
-                    totalBorrowedFiat:
-                      pool.totalBorrowedNative * prices[pool.chainId.toString()].value,
-                    totalSupplyBalanceFiat:
-                      pool.totalSupplyBalanceNative * prices[pool.chainId.toString()].value,
-                    totalBorrowBalanceFiat:
-                      pool.totalBorrowBalanceNative * prices[pool.chainId.toString()].value,
-                  };
-
-                  return adaptedFusePoolData;
-                })
-              );
-
-              return allPools;
-            } catch (e) {
-              console.error(e);
-
-              return undefined;
+              if (!visiblePools?.length) {
+                visiblePools = poolSort(visiblePools);
+              }
             }
+
+            chainPools[sdk.chainId] = visiblePools;
+            _allPools.push(...visiblePools);
+
+            const allPools: PoolData[] = await Promise.all(
+              _allPools.map((pool) => {
+                const assetsWithPrice: MarketData[] = [];
+                const { assets } = pool;
+
+                if (assets && assets.length !== 0) {
+                  assets.map((asset) => {
+                    assetsWithPrice.push({
+                      ...asset,
+                      supplyBalanceFiat:
+                        asset.supplyBalanceNative * prices[pool.chainId.toString()].value,
+                      borrowBalanceFiat:
+                        asset.borrowBalanceNative * prices[pool.chainId.toString()].value,
+                      totalSupplyFiat:
+                        asset.totalSupplyNative * prices[pool.chainId.toString()].value,
+                      totalBorrowFiat:
+                        asset.totalBorrowNative * prices[pool.chainId.toString()].value,
+                      liquidityFiat: asset.liquidityNative * prices[pool.chainId.toString()].value,
+                    });
+                  });
+                }
+                const adaptedFusePoolData: PoolData = {
+                  ...pool,
+                  assets: assetsWithPrice,
+                  totalLiquidityFiat:
+                    pool.totalLiquidityNative * prices[pool.chainId.toString()].value,
+                  totalAvailableLiquidityFiat:
+                    pool.totalAvailableLiquidityNative * prices[pool.chainId.toString()].value,
+                  totalSuppliedFiat:
+                    pool.totalSuppliedNative * prices[pool.chainId.toString()].value,
+                  totalBorrowedFiat:
+                    pool.totalBorrowedNative * prices[pool.chainId.toString()].value,
+                  totalSupplyBalanceFiat:
+                    pool.totalSupplyBalanceNative * prices[pool.chainId.toString()].value,
+                  totalBorrowBalanceFiat:
+                    pool.totalBorrowBalanceNative * prices[pool.chainId.toString()].value,
+                };
+
+                return adaptedFusePoolData;
+              })
+            );
+
+            return allPools;
           }
         },
         cacheTime: Infinity,
